Add bubble sort algorithm and print functions
This commit adds the bubble sort algorithm implementation in the file "0-bubble_sort.c". It also includes the implementation of the "print_array" and "print_list" functions in the files "print_array.c" and "print_list.c" respectively. These functions are used to print arrays and linked lists of integers.